requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cryptography==44.0.0
Django==5.1.4
django-bootstrap5==24.3
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.10
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
